Split tests into CRUD and errors suites
